requirements.txt
django==3.2.7
flake8==3.9.2
black==21.9b0
mypy==0.910
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Pillow==8.3.2
